Add style and finish game logic
Added style to the game board to make the page appear cleaner to a user.
Game board and button to create game are hid at start, and when a create
game it submitted the sign-in message is changed back to blank.
Some of these changes broke the game logic which had to be debugged.
Currently no known errors are appearing and game logic runs
successfully.