Fixed bug and changed styles
Discovered a problem that would display a particular player won when
they shouldn't have. The previous commit changed the game logic of the
gameDraw with an every function as it appeared that was the potential
problem. That change has been left in as it simplifies the code. The
problem part was in the winning combos arrays. The last combo was
defined as [2, 4, 7] instead of [2, 4, 6].  This made the issue hard to
spot initially. The correct combo has been replaced.
The style change rotated the board in a 3d field.